unpackbootimg: Add support for writing out hash type
With mkbootimg supporting different hash types, add the reverse
functionality to unpackbootimg, writing the hash type to a -hash file.

The algortithm for detecting the hash type isn't very clever. It assumes
that any ID which ends with 12 bytes of zeroes is sha1, and any hash
where none of the uint32_ts are zero is sha256. Anything else gives
"unknown".

This isn't infallible, but should work pretty reliably for the only two
hashes we know of. The chances of any hash having a 4-byte-aligned
section of zeroes 4 bytes long in it is pretty slim, and the chances of a
sha256 hash ending with 12 bytes of zeroes isn't even worth worrying
about (you can quote me on that when this breaks ;-).