unpackbootimg: add "second_signature" workaround support
Adapted from diff submitted by "Srinivasa R Kanduru" <skanduru@gmail.com>

... Thank you!

Turns out Google is also using this method to quietly sign their Pixel images, with no documentation on AOSP, however the bootloader isn't enforcing it like Xiaomi's appears to be.

Original unmkbootimg patch diff: https://pastebin.com/QwYZaeAx

--------------------------------------------------------------

"Hi osm0sis,

I was trying to split boot.img using unmkbootimg for my Xiaomi Mi MAX phone and found that it was always causing soft-brick by flashing the packed image. When I did a binary compare found what seemed like a second_image below ramdisk.

The header was showing the second_image size as '0' whereas the second_image did exist with a size of 1354 bytes.

unmkbootimg rightly ignores the second_image and mkbootimg fails to patch it. Net result is the image doesn't boot.

I made a small change to look for the second_image at the end of ramdisk in 'short ints' and if it exists try to calculate and populate the second_image_size. This ensures that the merged image does have the second_image.

Attached are the unmkbootimg.c diffs and the offending boot.img image.

With this workaround the image created by mkbootimg is identical to the original one except for 2 changes:
1. second_image_size is non-zero
2. SHA1 gets added at address 0x240 as inserted by mkbootimg.

I would like to have your opinion on this workaround since there would be lot of phones out there in the market with this issue.

Just saw your tool and it seems to have the same logic. Maybe this fix will work for unpackbootimg as well.

Thanks,
Srinivas"