Fix handling of extra-long boot image command-lines
The previous handling of command-lines equal to or longer than
EXTRA_ARGS_SIZE was broken.  Based on reading AOSP code it appears the
command-line area is nil-character terminated IF the command-line is
shorter than the space, but if the command-line exceeds that length there
is no nil-character.  Also appears the full length of the extra area can
be used, but I've left the limit 1 character below that to keep changes
smaller.